pyproject.toml
[tool.poetry]
name = "telegram-bot-sample"
version = "0.1.0"
description = ""
authors = ["manabreakmura <174273321+manabreakmura@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-decouple = "^3.8"
python-telegram-bot = "^21.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
